+++ src/helpers/json_parse.js
@@ -1,0 +1,9 @@
+const JsonBigInt = require('json-bigint')()
+
+export const parseJsonBigInt = (...args) => {
+  return JsonBigInt.parse(...args)
+}
+
+export const stringifyJsonBigInt = (...args) => {
+  return JsonBigInt.stringify(...args)
+}