+++ src/components/Snapshots/SnapshotsTable.vue
@@ -1,5 +1,5 @@
 <template>
-  <v-data-table :footer-props="{itemsPerPageOptions: DEFAULT_ITEMS_PER_PAGE}"
+  <v-data-table :footer-props="{itemsPerPageOptions: DEFAULT_ITEMS_PER_PAGE, showFirstLastPage: true}"
                 :headers="HEADERS"
                 :items="items"
                 :loading="loading"